+++ pipelines/06-redpanda-risingwave/dbt_project/models/intermediate/int_hourly_patterns.sql
@@ -15,10 +15,10 @@ hourly_agg as (
         is_weekend,
 
         count(*) as total_trips,
-        round(avg(trip_distance_miles), 2) as avg_distance,
-        round(avg(trip_duration_minutes), 2) as avg_duration_min,
-        round(sum(total_amount), 2) as total_revenue,
-        round(avg(total_amount), 2) as avg_revenue
+        round(avg(trip_distance_miles)::numeric, 2) as avg_distance,
+        round(avg(trip_duration_minutes)::numeric, 2) as avg_duration_min,
+        round(sum(total_amount)::numeric, 2) as total_revenue,
+        round(avg(total_amount)::numeric, 2) as avg_revenue
 
     from trip_metrics
     group by pickup_date, pickup_hour, pickup_day_of_week, is_weekend
